+++ user/src/main/java/com/user/filter/JwtFilter.java
@@ -1,5 +1,6 @@
 package com.user.filter;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.user.service.CustomUserDetailsService;
 import com.user.service.JwtService;
 import jakarta.servlet.FilterChain;
@@ -7,8 +8,12 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -33,7 +38,13 @@ public class JwtFilter extends OncePerRequestFilter {
 
     if (authHeader != null && authHeader.startsWith("Bearer ")) {
       token = authHeader.substring(7);
-      username = jwtService.extractUsername(token);
+      try {
+        username = jwtService.extractUsername(token);
+      } catch (Exception e) {
+        handleException(response, e);
+        // log
+        return;
+      }
     }
 
     if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
@@ -49,5 +60,18 @@ public class JwtFilter extends OncePerRequestFilter {
     }
 
     filterChain.doFilter(request, response);
+  }
+
+  private void handleException(HttpServletResponse response, Exception ex) throws IOException {
+    response.setStatus(HttpStatus.UNAUTHORIZED.value());
+    response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+
+    String[] errors = {"Token is expired or invalid"};
+    Map<String, Object> responseBody = new HashMap<>();
+    responseBody.put("message", "Unauthorized");
+    responseBody.put("errors", errors);
+
+    response.getWriter().write(new ObjectMapper().writeValueAsString(responseBody));
+    response.getWriter().flush();
   }
 }