+++ user/src/main/resources/db/migration/V2__Create_followers_table.sql
@@ -1,0 +1,6 @@
+CREATE TABLE followers (
+    follower_id UUID NOT NULL REFERENCES users(id) ON UPDATE CASCADE ON DELETE CASCADE,
+    influencer_id UUID NOT NULL REFERENCES users(id) ON UPDATE CASCADE ON DELETE CASCADE,
+    created_at TIMESTAMPTZ NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    PRIMARY KEY (follower_id, influencer_id)
+);